class and package shuffling
