+++ lib/Wikia/src/Domain/User/Preferences/GlobalPreference.php
@@ -1,22 +1,19 @@
 <?php
 
-namespace Wikia\Domain\User;
+namespace Wikia\Domain\User\Preferences;
 
 use Wikia\Util\Assert;
 
-class LocalPreference {
+class GlobalPreference {
 
 	private $name;
 	private $value;
-	private $wikiId;
 
-	function __construct( $name, $value, $wikiId ) {
+	function __construct( $name, $value ) {
 		Assert::true(!empty($name), "invalid preference name");
-		Assert::true(!empty($wikiId), "invalid wiki id");
 
 		$this->name = $name;
 		$this->value = $value;
-		$this->wikiId = $wikiId;
 	}
 
 	public function getName() {
@@ -25,10 +22,6 @@ class LocalPreference {
 
 	public function getValue() {
 		return $this->value;
-	}
-
-	public function getWikiId() {
-		return $this->wikiId;
 	}
 
 	public function setValue($value) {
