+++ lib/Wikia/src/Persistence/User/Preferences/PreferencePersistence.php
@@ -2,8 +2,7 @@
 
 namespace Wikia\Persistence\User\Preferences;
 
-use Wikia\Domain\User\GlobalPreference;
-use Wikia\Domain\User\Preferences;
+use Wikia\Domain\User\Preferences\UserPreferences;
 use Wikia\Service\PersistenceException;
 use Wikia\Service\UnauthorizedException;
 
@@ -13,18 +12,18 @@ interface PreferencePersistence {
 	 * Save the users preferences.
 	 *
 	 * @param int $userId
-	 * @param Preferences $preferences
+	 * @param UserPreferences $preferences
 	 * @return true success, false or exception otherwise
 	 * @throws PersistenceException
 	 * @throws UnauthorizedException
 	 */
-	public function save( $userId, Preferences $preferences );
+	public function save( $userId, UserPreferences $preferences );
 
 	/**
 	 * Get the users preferences.
 	 *
 	 * @param int $userId
-	 * @return Preferences
+	 * @return UserPreferences
 	 * @throws UnauthorizedException
 	 * @throws PersistenceException
 	 */
